split hooks to own files, update imports, fix tests
